refactor: valid 함수를 utils로 분리
