+++ src/controller/handleCountInput.js
@@ -1,9 +1,6 @@
-import { ERROR_MESSAGES, MIN_COUNT } from '../util/constants.js';
+import { ERROR_MESSAGES } from '../utils/constants.js';
+import { isValidCount } from '../utils/valid.js';
 import playRace from './playRace.js';
-
-function isValidCount(count) {
-  return +count >= MIN_COUNT;
-}
 
 export default function handleCountInput() {
   const { value } = document.querySelector('.count-input');
